CInt backed by long. CInt, CRange index support for string. Project Euler testing CRange to list fix. Dusting off cobwebbs.

+++ src/ExecutionTree/IntegerLiteralETNode.java
@@ -10,10 +10,10 @@ import Scanning.Token;
  */
 public class IntegerLiteralETNode extends ETNode {
 
-    public int value;
+    public long value;
 
     public IntegerLiteralETNode(Token token) {
-        this.value = Integer.parseInt(token.sequence);
+        this.value = Long.parseLong(token.sequence);
     }
 
     public IntegerLiteralETNode(int value) {
